fix random compiler warnings

+++ core/util/communications/referee/ref_ui.cpp
@@ -1,4 +1,5 @@
 #include "ref_serial.h"
+#include <cstddef>
 
 void Referee::ui_graph_characters(int operation_type, string str, int x, int y, char name){
     ext_student_interactive_header_data_character_t custom_character_draw;
@@ -7,7 +8,7 @@ void Referee::ui_graph_characters(int operation_type, string str, int x, int y, 
         custom_character_draw.sender_ID=get_robot_id();//发送者ID，机器人对应ID
         custom_character_draw.receiver_ID=get_robot_id()+0x0100;//接收者ID，操作手客户端ID
         {
-            for(int i=0; i<str.length(); i++){
+            for(size_t i=0; i<str.length(); i++){
                 custom_character_draw.graphic_custom.data[i]=str[i];
             }
             for(int i=str.length(); i<30; i++){ // have to fill in the rest of the char array
